transformations between int8 and uint8

+++ modules/s2/redshift/sql/GET_RESOLUTION.sql
@@ -3,13 +3,13 @@
 ----------------------------
 
 CREATE OR REPLACE FUNCTION @@RS_PREFIX@@s2.GET_RESOLUTION(
-    id BIGINT
+    id INT8
 )
 RETURNS INTEGER 
 IMMUTABLE
 AS $$
     from @@RS_PREFIX@@s2Lib import get_resolution
     
-    return get_resolution(long(id))
+    return get_resolution(id)
     
 $$ LANGUAGE plpythonu;
